requirements.txt
dicttoxml
xmltodict
beautifulsoup4
lxml
pyyaml
rec-avro
avro
dataclasses-avroschema
typing_extensions